+++ CHANGES.md
@@ -15,6 +15,7 @@
 + core: Add `set_tooltip` method to `RelmWidgetExt`
 + core: Add `main_adw_application` method to retrieve the `adw::Application` when the libadwaita feature is enabled
 + macros: Add `skip_macro` option for watch and track attributes to skip their initialization 
++ macros: Improve error messages for non-identifier parameter patterns
 + examples: Example for using relm4-icons 
 
 ### Changed